changed component value datatype to double, added converting to rot and pot, master can send boundsinfo to a node and a node can receive it and assign the bounds to a component
Problem with packet ids as I am sending a struct for the bounds which doesnt have a id... Now the receiver function is just an if else id == reset and else if the struct. The problem is that you can still read the id from the struct which can give some unwanted results. I will fix this next.

+++ ArduinoFiles/Components/Button/Button.h
@@ -13,7 +13,7 @@ class Button: public Component
 public:
     Button() {}
     Button( int pin, uint8_t id);
-    uint16_t getValue();
+    double getValue();
 
 private:
     int _pin;
